+++ src/main/java/com/restaurante01/api_restaurante/produto/dto/entrada/ProdutoCreateDTO.java
@@ -2,6 +2,8 @@ package com.restaurante01.api_restaurante.produto.dto.entrada;
 
 import jakarta.validation.constraints.*;
 
+import java.math.BigDecimal;
+
 public class ProdutoCreateDTO {
     @NotBlank(message = "Nome não deve ser vazio")
     private String nome;
@@ -9,13 +11,13 @@ public class ProdutoCreateDTO {
     private String descricao;
     @NotNull(message = "Preço deve ser informado")
     @PositiveOrZero(message = "Preço deve ser zero ou positivo")
-    private Double preco;
+    private BigDecimal preco;
     @NotNull(message = "Quantidade deve ser zero ou positivo")
     @Min(value = 0, message = "Quantidade minima deve ser zero")
     private Long quantidadeAtual;
     private Boolean disponibilidade;
 
-    public ProdutoCreateDTO(String nome, String descricao, Double preco, Long quantidadeAtual, Boolean disponibilidade) {
+    public ProdutoCreateDTO(String nome, String descricao, BigDecimal preco, Long quantidadeAtual, Boolean disponibilidade) {
         this.nome = nome;
         this.descricao = descricao;
         this.preco = preco;
@@ -35,7 +37,7 @@ public class ProdutoCreateDTO {
         return disponibilidade;
     }
 
-    public Double getPreco() {
+    public BigDecimal getPreco() {
         return preco;
     }
