+++ src/main/java/com/restaurante01/api_restaurante/produto/entity/Produto.java
@@ -5,6 +5,8 @@ import jakarta.persistence.*;
 import jakarta.validation.constraints.*;
 import lombok.EqualsAndHashCode;
 
+import java.math.BigDecimal;
+
 @Entity
 @Table(name = "produto")
 @EqualsAndHashCode(callSuper = true)
@@ -21,7 +23,7 @@ public class Produto extends Auditable {
 
     @NotNull(message = "Quantidade deve ser zero ou positivo")
     @PositiveOrZero(message = "preço deve ser positivo")
-    private Double preco;
+    private BigDecimal preco;
 
 
     @Min(value = 0, message = "Quantidade minima deve ser zero")
@@ -32,7 +34,7 @@ public class Produto extends Auditable {
     public Produto(){
 
     }
-    public Produto(Long id,String nome, String descricao, Double preco, Long quantidadeAtual, Boolean disponibilidade) {
+    public Produto(Long id,String nome, String descricao, BigDecimal preco, Long quantidadeAtual, Boolean disponibilidade) {
         this.id = id;
         this.nome = nome;
         this.descricao = descricao;
@@ -65,11 +67,11 @@ public class Produto extends Auditable {
         this.descricao = descricao;
     }
 
-    public Double getPreco() {
+    public BigDecimal getPreco() {
         return preco;
     }
 
-    public void setPreco(Double preco) {
+    public void setPreco(BigDecimal preco) {
         this.preco = preco;
     }
